refactor: move features in core into their own folders
