+++ ng-client/src/app/core/login/login.component.scss
@@ -1,4 +1,4 @@
-@import "../app.colors";
+@import "../../app.colors";
 
 .alert-container {
   height: 60px;
