+++ ng-client/src/app/core/navigation/navigation.component.scss
