+++ ng-client/src/app/core/navigation/navigation.component.spec.ts
@@ -8,9 +8,9 @@ import { Router } from '@angular/router';
 
 import { Observable } from 'rxjs/Observable';
 
-import { IdentityService, LocalStorage, AUTH_TOKEN_NAME } from '../core';
+import { IdentityService, LocalStorage, AUTH_TOKEN_NAME } from '../';
 
-import { User, Name } from '../core/user';
+import { User, Name } from '../user';
 
 import { NavigationComponent } from './navigation.component';
